refactot Navigate -> NavBar
